backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.11.1
eventlet==0.33.3
openai==1.58.1
numpy==1.26.3
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
requests==2.31.0
Pillow==10.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
